backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic[email]>=2.0.0
httpx>=0.25.0
python-dotenv>=1.0.0
sqlalchemy>=2.0.0
aiogram>=3.0.0
